optimized build for lazy loading
