+++ src/index.tsx
@@ -4,7 +4,6 @@ import './css'
 export * from './Components/Map'
 export * from './Components/AppShell'
 export * from './Components/Auth'
-export * from './Components/Profile'
 export * from './Components/Gaming'
 export * from './Components/Templates'
 export * from './Components/Input'
